lib/bootstrap.sh: change load_disk_fail to load_disk.fail
This is easily bisect load disk issue along with the assistence by providing
load_disk.pass to be recorded in last_state. The side effect is it generates
last_state file with load_disk.pass which is confusing. Further refinement is
needed to record this as part of whole test output.

Signed-off-by: Philip Li <philip.li@intel.com>