include/perf-profile: Disable RT_GROUP_SCHED
When perf-profile runs from a child cgroup and cpu controller is
enabled in cgroup v2, perf will not be able to switch itself to realtime
policy due to its group's rt_bandwidth.rt_runtime is 0 and there is no
interface to set a positive value for this field under cgroup v2.

Disable this config to make perf-profile work under cgroupv2 with cpu
controller enabled, but only for kernels that are not built yet due
to suggest_kconfig.

Also note that config-5.14.0-171.el9.x86_64 doesn't enable this config
either.

Suggested-by: Philip Li <philip.li@intel.com>
Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>