tests/makepkg: stop packing when installing packages failed
when installing dev packages failed, some libraries and corresponding
hearder files may not be installed, so the build of test suite may be
lack of some libs, and causes some testcases not able to run.

stop the packing process if the installation of dev packages failed, so
we can recognize above issue in pack time instead of run time.

Signed-off-by: Yujie Liu <yujie.liu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>