repo/linux/ceph-client: update git_am_branch
Apply patch to testing branch may be a good choose:
Re: [PATCH] ceph: fix memory leak in mount error path when using test_dummy_encryption
-------------
> we also appreciate that if developers can tell us the right branch
> to improve the bot when applied to wrong place.

Yeah, I guess that in general the bot is picking the right branch for
ceph.  In this case, the patch was for the fscrypt development branch, so
my mistake for not using '--base'.
-------------

Link: https://lore.kernel.org/all/Y2T1ocj1xirSc2Lf@suse.de/
Signed-off-by: Chen Rong <rong.a.chen@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>