tests/schbench: make it possible to set intervaltime
Intervaltime is used to determine how often schbench print current
latency stats and the printed stats can affect stat/schbench calculating
latency average and stddev.

When some large runtime value is used, the purpose is to have the
workload itself to stablize and deliver some stable latency values for
the entire $runtime time. But if every 10s(default interval) a latency
stat is printed, that can affect stat/schbench to see too many changing
values and that can affect schbench's stddev.

Make it possible to specify intervaltime so users who don't want to see
many latency stats can set intervaltime to be the same as runtime to get
a single latency stat at the end of the test.

Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>