include/kernel-selftests: update RC_CORE setting
RC_CORE was updated to y by below commit for bpf selftests, but
BPF_LIRC_MODE2 has been moved to kernel-selftests-bpf now, so RC_CORE is
not necessarily to be y anymore, and we just allow it to be either y or m.

  f63762294 include/kernel-selftests: fix LKP WARN miss config of bpf/config

update corresponding comments to remove unnecessary info.

Signed-off-by: Yujie Liu <yujie.liu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>