stats/hwsim: add hwsim build failure
spotted a hwsim build failure between v5.14 and v5.15-rc1. let stater
handle this so we can bisect it.

it is easy to judge whether make is fail since there is clear error
log, but not very accurate to judge whether make is pass. here we
anticipate that make is pass if "./run-tests.py" is called, which means
"./build.sh" and "./start.sh" have no issue.

Signed-off-by: Yujie Liu <yujie.liu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>