monitors: Better align the runtime and delay of perf-profile/perf-sched/perf-c2c
Likely, several kinds of perf monitors which could be running at the
same time: perf-profile, perf-sched and perf-c2c. There are already
some consideration for it, and still a better alignment is needed.

Basically, if runtime/testtime is given by the benchmark, the delay
for 3 monitors will be set to 1/4, 1/2 and 3/4 of them. Otherwise
the delay will be separated to 30/60/100 seconds.

Signed-off-by: Feng Tang <feng.tang@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>