stats/perf-c2c: add HITM.total metric
$ perf c2c --help

DESCRIPTION
       C2C stands for Cache To Cache.

       The perf c2c tool provides means for Shared Data C2C/HITM analysis. It allows you to track
       down the cacheline contentions.

       The c2c tool provide means to record this data and report back access details for
       cachelines with highest contention - highest number of HITM accesses.

False sharing occurs when a processor core on a Symmetric Multi
Processing (SMP) system modifies data items on the same cache line that
is in use by other processors to access other data items that are not
being shared between the processors.

This initial modification requires that the other processors using the
cache line invalidate their copy and request an updated one despite the
processors not needing, or even necessarily having access to, an updated
version of the modified data item.

You can use the perf c2c command to detect false sharing.

during the enabling phase, we use the sum of HITM.local and HITM.remote
as an indicator of potential false sharing issue. this can be changed
after we gather more data.

Link: https://access.redhat.com/documentation/en-us/red_hat_enterprise_linux/8/html/monitoring_and_managing_system_status_and_performance/detecting-false-sharing_monitoring-and-managing-system-status-and-performance
Signed-off-by: Yujie Liu <yujie.liu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>