stats/perf-stat: capture the expected events accurately
perf stat command can accept an events list in -e option to specify the
target events to be captured.

there are some cases that the characters in perf stat output are
completely out of order, so the stats are wrong after parsing.

579.1799 96735,1676,,pmigrations,6,nodf3767,48, a6017.998,/sec
582.1854 96555m6,,cpu-3nstrB-ions,6,nodf3767,48,.00,26.259,/sec

it should be "cpu-migrations" and "instructions", don't know why it
becomes a mess.

it would be better to get the events list from perf command, and
explicitly check it when reading each line to make sure it is the
expected event we want to capture but not some out-of-order wrong words.

add fake perf command in old spec to keep supporting it.

== fix issue ==

time: 579.1799
i.pmigrations: 1671.8511175481513
time: 582.1854
i.6: 0.0
Traceback (most recent call last):
	4: from /lkp/lkp/src/stats/perf-stat:277:in `<main>'
	3: from /lkp/lkp/src/stats/perf-stat:172:in `parse'
	2: from /lkp/lkp/src/stats/perf-stat:172:in `each_line'
	1: from /lkp/lkp/src/stats/perf-stat:193:in `block in parse'
/lkp/lkp/src/stats/perf-stat:222:in `parse_per_socket_mode': undefined method `[]' for nil:NilClass (NoMethodError)

Signed-off-by: Yujie Liu <yujie.liu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>