include/category/benchmark: Do not disable CONFIG_SLUB_DEBUG
SLUB_DEBUG is a useful and lightweight debug option, which is
a 'default y' in kernel config, and enabled by CentOS 9 and Ubuntu etc.

Also if we force to disable it, when debugging some slub related
regression, it's not convenient as we can't enable it in our custom
test-build config.

Signed-off-by: Feng Tang <feng.tang@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>