programs/kernel-selftests/parse: fix a net stater pattern
the original pattern is too general and can even matches some
irrelevant log such as:

> line = '#  mz/getopts: libnet_init() failed (socket: Invalid argument) Invalid command line parameters!'
> puts 'matched' if line =~ /#  ([a-zA-Z].*)(ok|fail)/
matched

the original pattern ([a-zA-Z].*) doesn't make much sense. [a-zA-Z] only
matches one charater, and the remaning .* can match anything. there is
no example in the comment to prove this pattern is valid.

fix this pattern to be more strict. not sure if this may impact other
cases, but currently all the stats specs can still pass.

$ rake spec spec=stats
/usr/bin/ruby3.0 -I/var/lib/gems/3.0.0/gems/rspec-core-3.11.0/lib:/var/lib/gems/3.0.0/gems/rspec-support-3.11.0/lib /var/lib/gems/3.0.0/gems/rspec-core-3.11.0/exe/rspec --pattern spec/\*\*\{,/\*/\*\*\}/stats_spec.rb
..................................................................................................................................................................................................................................................................................................................................................................................................

Finished in 1 minute 1.06 seconds (files took 0.14338 seconds to load)
386 examples, 0 failures

Signed-off-by: Yujie Liu <yujie.liu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>