repo/linux/chrome-os: Drop chromeos-4.14 and chromeos-4.19
chromeos-4.14 and chromeos-4.19 are close to EOL. While identifying and
fixing real bugs in those branches would still be important, 0-day is
creating too much noise which distracts from real problems. Since there
does not appear to be a means to selectively tell 0-day to not check for
documentation mismatches or other cosmetic issues in older branches (or
anywhere, really), drop coverage of close-to-eol branches to reduce the
noise.

Signed-off-by: Guenter Roeck <groeck@chromium.org>