filters/need_kconfig.rb: fix unexpected unsatisfied constraints
If a kconfig can be either y or m, the kconfig requirement is written as "- CONFIG_NAME".
Such kconfig doesn't follow the constraints defined in etc/kconfigs.yaml, such as

	["CRYPTO_OFB"] has not been compiled by this kernel (v6.8-rc2 based)

This is caused by code bug in need_kconfig.rb which missed to handle the code flow
of such kconfig to query etc/kconfigs.yaml.

Signed-off-by: Philip Li <philip.li@intel.com>