kernel-selftests: rename atomic_test to test_atomic
The path params are sorted alphabetically. The following case has three
params:

kernel-selftests:
  group:
    - net
  test:
    - fcnal-test.sh
  atomic_test:
    - ipv4_ping

They will be sorted as "atomic_test", "group", "test", so the
final rt path will be:

/result/kernel-selftests/ipv4_ping-net-fcnal-test.sh/...

This brings difficulty to determine which group it belongs to. We hope
to put "group" param at the first position, so rename atomic_test to
test_atomic to put it at the last position. The rt path will be:

/result/kernel-selftests/net-fcnal-test.sh-ipv4_ping/...

Signed-off-by: Yujie Liu <yujie.liu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>