jobs/phoronix-test-suite*: disable some opencl tests
These tests raise a similar "OpenCL support seems to be unavailable"
issue.

* They check the output of "clinfo" command to get the number of
available platforms, which is 0 in our env.

* We tried to install mesa-opencl-icd as suggested by the search result
on the Internet. After installing it, the number of platforms is not 0
now, but still raises error when opening the device.

OpenCL Platform 0: Mesa
OpenCL Platform 1: Mesa/X.org
Failed to open OpenCL context

* If mesa-opencl-icd is installed, other opencl benchmarks (such as
financebench) will fail, so we must not install mesa-opencl-icd to avoid
breaking other ones.

* There is no good idea to fix them, because other opencl benchmarks can
work well, which means the issue may not be related with opencl, but may
be related with the particular benchmarks themselves.

Signed-off-by: Yujie Liu <yujie.liu@intel.com>
Signed-off-by: Philip Li <philip.li@intel.com>