distro/adaptation-pkg: map turbostat program to distro pkg
This is to avoid local makepkg of turbostat (such as during reproducing issue)
and use pre-installed turbostat instead.

The turbostat built from kernel is useful to run latest platform, but most of
them this is not required to use the latest turbostat.

On the other side, this patch blocks the use of latest turbostat, which need
be resolved in later logic to allow user to choose which way is preferred.

Link: https://lore.kernel.org/oe-lkp/2145949.1714121817@warthog.procyon.org.uk/
Signed-off-by: Philip Li <philip.li@intel.com>